email_agent/src/requirements.txt
# langgraph==0.2.14
# google-api-python-client==2.137.0
# google-auth==2.35.0
# google-auth-oauthlib==1.2.1
# openai==1.42.0
# google-auth-httplib2==0.2.0

langgraph==0.6.4
google-api-python-client==2.137.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
openai==1.42.0
boto3==1.28.0
